requirements.txt
fastapi
uvicorn
argostranslate
python-dotenv
